signal/requirements.txt
redis==5.0.8
alpaca-py==0.26.0
transformers==4.44.2
torch==2.4.1
spacy==3.7.6
python-dotenv==1.0.1
vaderSentiment==3.3.2